chore: refactor construction visual
