+++ src/apps/Construction/Components/ConstructionVisual/Components/Legend/index.ts
