+++ src/apps/Construction/Components/ConstructionVisual/Components/Legend/styles.tsx
