+++ src/apps/Construction/Components/ConstructionVisual/ConstructionVisual.tsx
@@ -3,11 +3,12 @@ import { ConstructionGraphProps } from './Types/constructionVisualOptions';
 
 export function ConstructionVisual<T extends unknown>({
     data,
-    options,
+    defaultTime,
+    title,
 }: ConstructionGraphProps<T>): JSX.Element | null {
     if (data.length <= 0) {
         return null;
     }
 
-    return <Graph data={data} options={options} />;
+    return <Graph data={data} defaultTime={defaultTime} title={title} />;
 }
