+++ src/apps/Construction/Components/ConstructionVisual/Styles/Styles.ts
