+++ src/apps/Construction/Components/ConstructionVisual/Types/constructionVisualOptions.ts
@@ -1,0 +1,7 @@
+import { TimeDimension } from '@equinor/Diagrams';
+
+export interface ConstructionGraphProps<T> {
+    data: T[];
+    title?: string;
+    defaultTime?: TimeDimension;
+}
