+++ src/apps/Construction/Components/ConstructionVisual/Types/index.ts
@@ -1,0 +1,2 @@
+export * from './series';
+export * from './constructionVisualOptions';
