+++ src/apps/Construction/Components/ConstructionVisual/Utils/colorGenerator.ts
