+++ src/apps/Construction/Components/ConstructionVisual/Utils/htmlLegendPlugin.tsx
