+++ src/apps/Construction/Components/index.ts
@@ -1,2 +1,4 @@
 export * from './CriticalWoTable';
 export * from './DetailsPage';
+export * from './ConstructionVisual';
+export * from './SidesheetContent';
