+++ src/packages/Diagrams/src/Components/index.ts
@@ -1,3 +1,2 @@
 export * from './AnalyticsView';
 export * from './GetCharts';
-export * from './SidesheetContent';
