+++ src/packages/Diagrams/src/Types/index.ts
@@ -1,3 +1,4 @@
 export * from './analyticsOptions';
 export * from './cumulativeSeries';
 export * from './cumulativeSeriesOptions';
+export * from './timeDimension';
