+++ src/packages/Diagrams/src/Types/timeDimension.ts
@@ -1,0 +1,1 @@
+export type TimeDimension = 'month' | 'week' | 'year' | 'quarter';
