+++ src/packages/Diagrams/src/Visuals/index.ts
@@ -3,4 +3,3 @@ export * from './ControlledTimeVisual';
 export * from './LineVisual';
 export * from './TableVisual';
 export * from './TimeVisual';
-export * from './ConstructionVisual';
